requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
passlib>=1.7.4
python-dotenv>=1.0.0

# Validation
pydantic[email]>=2.0.0
email-validator>=2.0.0

# Templating
jinja2>=3.1.0

# Scheduling
apscheduler>=3.10.0

# HTTP Client (for testing)
requests>=2.28.0
httpx>=0.24.0
